Implementação do Map para a versão Lambari
